README & config.json relocate
