+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ayaka",
   "version": "1.0.1",
-  "description": "Kamisato Ayaka from Genshin Impact with openai-node backend",
+  "description": "Discord chatbot with openai-node backend",
   "main": "dist/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
