+++ src/events/ready.ts
@@ -1,5 +1,4 @@
-// @ts-ignore
-import { on_ready_msg } from "../../config.json";
+import { on_ready_msg } from "../config.json";
 const { Events } = require("discord.js");
 
 module.exports = {
