+++ src/utils/stringManipulation.ts
@@ -1,5 +1,4 @@
-// @ts-ignore
-import { separator } from "../../config.json";
+import { separator } from "../config.json";
 
 export function lowerFirstChar(str: string) {
   return str.charAt(0).toLowerCase() + str.substr(1);
